Move repo lifecycle operations from HubSupervisor into RepoManager
